+++ src/test/kotlin/krangl/test/ColumnTests.kt
@@ -1,18 +1,59 @@
 package krangl.test
 
-import io.kotlintest.specs.FlatSpec
+import io.kotlintest.TestBase
+import io.kotlintest.TestFailedException
+import io.kotlintest.matchers.Matchers
+import krangl.AnyCol
+import krangl.BooleanCol
+import krangl.DoubleCol
 import krangl.IntCol
+import org.junit.Test
 
 /**
  * @author Holger Brandl
  */
-class ColumnTests : FlatSpec() { init {
+class ColumnTests : Matchers {
 
-    "it" should "do correct column arithmetics" {
+    @Test
+    fun `it should do correct column arithmetics`() {
 
         (IntCol("", listOf(3)) + 3)[0] shouldBe 6
         (IntCol("", listOf(3)) + 3.0)[0] shouldBe 6.0
         (IntCol("", listOf(3)) + "foo")[0] shouldBe "3foo"
     }
+
+
+    @Test
+    fun `allow to negate and invert columns`() {
+
+        (!BooleanCol("foo", listOf(false, true)))[0] shouldBe true
+
+        (-IntCol("foo", listOf(1, 2)))[1] shouldBe -2
+        (-DoubleCol("foo", listOf(1.2, 2.0)))[1] shouldBe -2.0
+
+
+        shouldThrow<UnsupportedOperationException> { (-BooleanCol("foo", listOf(true))) }
+        shouldThrow<UnsupportedOperationException> { (!IntCol("foo", listOf(1))) }
+//
+        shouldThrow<UnsupportedOperationException> { (!AnyCol("foo", listOf(1))) }
+        shouldThrow<UnsupportedOperationException> { (-AnyCol("foo", listOf(1))) }
+
+    }
 }
+
+
+internal inline fun <reified T> shouldThrow(thunk: () -> Any): T {
+    val e = try {
+        thunk()
+        null
+    } catch (e: Exception) {
+        e
+    }
+
+    if (e == null)
+        throw TestFailedException("Expected exception ${T::class.qualifiedName} but no exception was thrown")
+    else if (e.javaClass.name != T::class.qualifiedName)
+        throw TestFailedException("Expected exception ${T::class.qualifiedName} but ${e.javaClass.name} was thrown")
+    else
+        return e as T
 }